refactored graph
